:zap: chore: abstract using controllers
